backend/requirements.txt
fastapi
uvicorn
python-dotenv
pydantic[email]
requests
pyyaml
langchain
langchain-community
langchain-groq
langchain-openai
langchain-google-genai
langchain-google-community[places]
langchain-tavily
langgraph
motor
pymongo
bcrypt
python-jose[cryptography]
